traffic-sender/requirements.txt
requests
python-consul2
prometheus-client
